Elements overview

Element API

Five.get_fields — Method

get_fields(::AbstractElement)

Returns a list with fields of element, e.g :u, :T, :θ

Ferrite.getnquadpoints — Function

getnquadpoints(::AbstractElement)

Returns the number of quadrature points in the element Used for allocating arrays for the material states

Five.getquadraturerule — Function

getquadraturerule(::AbstractElement)

Returns the number of quadrature rule used by the element Used for postprocessing and L2Projection

Ferrite.ndofs — Function

ndofs(::AbstractElement)

Returns the number of dofs in the element Used for pre-allocating element matrices, internal force matrices etc.

Ferrite.getcelltype — Function

getcelltype(::AbstractElement)

Returns the celltype (Ferrite.AbstractCell) of the element

Five.has_constant_massmatrix — Function

Returns true if the massmatrix is constant or not

Five.integrate_forcevector! — Function

Returns the internal forcevector of element

Five.integrate_forcevector_and_stiffnessmatrix! — Function

Returns the internal forcevector and the element stiffness matrix of element

Five.integrate_massmatrix! — Function

Returns the mass matrix of the element

Five.integrate_dissipation! — Function

Returns the dissipation and the gradient of the dissipation (used in LocalDissipationSolver)