Essentials

Five.StateVariables — Type

StateVariables(T::Type, ndofs::Int)

Contains all information about the state of system (displacements, material damage etc).

Field variables:

d: displacements

v: velocityes

a: accelerations

t: current time

λ: current loading factor (fᵉ = λ*f̂, used in arc-length solver)

L: Current step length (used in arc-length solvers)

Δd, Δv, Δa, Δt, Δλ, ΔL - Difference between current and previous timestep of variables above

system_arrays: Instance of SystemArrays

partstates: Contains PartState, one for each cell

prev_partstates: Contains PartState from previous timesteps, one for each cell

step: Number of steps taken up until this point

Possible changes Remove all Δ-variables, and require two states instead, (previous and current)

Five.SystemArrays — Type

SystemArrays(T::Type, ndofs::Int)

Contains global arrays, such as internal force vector and stiffness matrix.

Five.GlobalData — Type

GlobalData{dim,T,DH<:AbstractDofHandler}

Contains all information about the problem being solved, e.g Forces, Boundary conditions, Parts