Crisfield arc-length solver

This solver is based on the paper from 1981 by Crisfield (A fast incremental ...)

The arc-length constraint is defined as

\[ \varphi(\boldsymbol a, \lambda) = \Delta \boldsymbol a \Delta \boldsymbol a^T + \Delta\lambda^2 - \Delta L^2 = 0\]

where $\Delta L$ is a user-defined parameter limiting the step size.

Five.ArcLengthSolver — Type

ArcLengthSolver{T}

Arguments: - Δλ0: Initial size of load parameter - λ_max: - λ_min: - ψ: - maxsteps: Maximum number of steps - max_residual: Maximum value of the residual before aborting a step - optitr: - maxitr: - maxitr_first_step: - finish_criterion: Function