Solver overview

All solvers follow a similar pattern. Given a state::StateVariables, the solver tries to advance the state to the next timestep/loadstep. The step size is determined by the solver itself based on its input parameters. If the current step diverges, most solvers tries to modify (e.g. decrease) the step size in order to hopefully achieve convergence more easily.

The functions a solver (<: AbstractSolver) should implement are presented below.

Five.step! — Function

step!(solver::AbstractSolver, state::StateVariables, globaldata::GlobalData)

Given the current state of a system, the solver advances to the next state (mutating the input argument state). The next state (and step size) is determined by the parameters of the solver.

Base.isdone — Function

isdone(itr, state...) -> Union{Bool, Missing}

This function provides a fast-path hint for iterator completion. This is useful for mutable iterators that want to avoid having elements consumed, if they are not going to be exposed to the user (e.g. to check for done-ness in isempty or zip). Mutable iterators that want to opt into this feature should define an isdone method that returns true/false depending on whether the iterator is done or not. Stateless iterators need not implement this function. If the result is missing, callers may go ahead and compute iterate(x, state...) === nothing to compute a definite answer.

isdone(solver::AbstractSolver, state::StateVariables, globaldata::GlobalData)

Determines if the state has reached the termination criteria of the solver

Five.should_abort — Function

should_abort(solver::AbstractSolver, state::StateVariables, globaldata::GlobalData)

Determines if the solver should abort simulation (for example if it is not able to converge)